refactor: separate concerns & rename plugins
